Redesign von instruction set
I have complete reworked the instruction set of the ByteMachine. It is
now reduced to a bare minimum with still being usable. As a proof of
concept, I have implemented a MD5 checksum calculator which runs in my
new ByteMachine VM.
I think with this new architecture the consumption of LUTs in an FPGA
should be quite low,, using ram blocks even for the operand stack and
return stack.